+++ Assets/Scripts/RapidSim/Networks/WrappedArray.cs
@@ -5,6 +5,6 @@ namespace RapidSim.Networks
     [Serializable]
     public struct WrappedArray
     {
-        public float[] data;
+        public double[] data;
     }
 }
